feat(setup): Add all components needed to write a model folder
This commit adds all of the components needed to create radiance models and export them to the honeybee-radiance folder.

It also adds a bunch of sample files and 3 of them have a hacked-together component to export things to the honeybee radiance folder.  This way, @mostapharoudsari can test the code to ensure it's working as expected.