fix(recipe): Ensure recipes on the same model don't overwrite each other
This fix ensures that recipes run on the same model do not overwrite each other.  The model folder will get overwritten to ensure that any updates to the geometry from the Grasshopper interface are accounted for in any new recipe that is run.  But the results will not get overwritten.