fix(modifiers): Make inputs for translucent modifier more intuitive
I've come to realize that I have been using the Trans material wrong for all of these years because the properties in Radiance are not what they initially seem. And they are definitely not what you would get from any product manufacturer. So I am changing the component so that we can actually construct the material from transmittance and reflectance.